LayoutTests/imported/w3c/web-platform-tests/media-source/mediasource-addsourcebuffer-mode-expected.txt


PASS addSourceBuffer() sets SourceBuffer.mode to 'segments' when the generate timestamps flag is false
PASS addSourceBuffer() sets SourceBuffer.mode to 'sequence' when the generate timestamps flag is true
